refactor(vmm): split irq chip
