+++ virt/arcbox-vmm/src/irq/chip.rs
@@ -1,201 +1,13 @@
-//! Interrupt controller management.
-//!
-//! This module provides the IRQ chip abstraction for managing interrupts,
-//! including GSI mapping, trigger modes, and interrupt coalescing.
-
 use std::collections::HashMap;
-use std::sync::atomic::{AtomicBool, AtomicU32, AtomicU64, Ordering};
+use std::sync::atomic::{AtomicU32, Ordering};
 use std::sync::{Arc, Mutex, RwLock};
-use std::time::{Duration, Instant};
 
 use crate::error::Result;
 
-/// IRQ number type.
-pub type Irq = u32;
-
-/// Global System Interrupt number type.
-pub type Gsi = u32;
-
-/// Maximum number of IRQs.
-pub const MAX_IRQS: u32 = 256;
-
-/// Maximum number of GSIs (typically matches IOAPIC entries + legacy PICs).
-pub const MAX_GSIS: u32 = 24;
-
-/// Interrupt trigger mode.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum TriggerMode {
-    /// Edge-triggered: interrupt is signaled on level transition.
-    /// Device asserts then deasserts the IRQ line.
-    #[default]
-    Edge,
-    /// Level-triggered: interrupt remains asserted until acknowledged.
-    /// Device keeps line asserted until serviced.
-    Level,
-}
-
-/// IRQ configuration for a single interrupt line.
-#[derive(Debug, Clone)]
-pub struct IrqConfig {
-    /// The GSI this IRQ is mapped to.
-    pub gsi: Gsi,
-    /// Trigger mode for this IRQ.
-    pub trigger_mode: TriggerMode,
-    /// Whether this IRQ is currently asserted (for level-triggered).
-    pub asserted: bool,
-}
-
-/// Callback type for triggering interrupts on the hypervisor.
-///
-/// The callback receives (gsi, level) where level is true for assert.
-pub type IrqTriggerCallback = Box<dyn Fn(Gsi, bool) -> Result<()> + Send + Sync>;
-
-/// Configuration for timer-based interrupt coalescing.
-///
-/// Trades a small latency increase for significant wakeup reduction.
-/// When an interrupt fires, a coalescing window opens; additional interrupts
-/// within the window are accumulated and delivered as a single notification
-/// when the window expires or the count threshold is reached.
-#[derive(Debug, Clone)]
-pub struct CoalescingConfig {
-    /// Maximum delay before delivering a pending interrupt.
-    pub max_delay: Duration,
-    /// Force delivery after this many pending interrupts.
-    pub max_coalesce_count: u32,
-    /// Whether coalescing is enabled.
-    pub enabled: bool,
-}
-
-impl Default for CoalescingConfig {
-    fn default() -> Self {
-        Self {
-            max_delay: Duration::from_micros(50),
-            max_coalesce_count: 64,
-            enabled: true,
-        }
-    }
-}
-
-impl CoalescingConfig {
-    /// Preset for virtio-net: moderate latency tolerance.
-    #[must_use]
-    pub fn for_net() -> Self {
-        Self {
-            max_delay: Duration::from_micros(50),
-            max_coalesce_count: 64,
-            enabled: true,
-        }
-    }
-
-    /// Preset for virtio-blk: lower latency tolerance.
-    #[must_use]
-    pub fn for_block() -> Self {
-        Self {
-            max_delay: Duration::from_micros(25),
-            max_coalesce_count: 32,
-            enabled: true,
-        }
-    }
-
-    /// Preset for virtio-fs: batches well.
-    #[must_use]
-    pub fn for_fs() -> Self {
-        Self {
-            max_delay: Duration::from_micros(50),
-            max_coalesce_count: 64,
-            enabled: true,
-        }
-    }
-
-    /// Preset for virtio-vsock: control plane, latency insensitive.
-    #[must_use]
-    pub fn for_vsock() -> Self {
-        Self {
-            max_delay: Duration::from_micros(100),
-            max_coalesce_count: 128,
-            enabled: true,
-        }
-    }
-
-    /// Disabled coalescing — pass through immediately.
-    #[must_use]
-    pub fn disabled() -> Self {
-        Self {
-            enabled: false,
-            ..Self::default()
-        }
-    }
-}
-
-/// Per-IRQ coalescing state.
-///
-/// Tracks pending interrupt count and timer state for a single IRQ line.
-pub struct CoalescingState {
-    /// Number of interrupts accumulated in the current window.
-    pub pending_count: AtomicU32,
-    /// Whether the coalescing timer is armed.
-    pub timer_armed: AtomicBool,
-    /// When the timer was armed (for expiry check).
-    pub last_armed: Mutex<Option<Instant>>,
-    /// Configuration for this IRQ line.
-    pub config: CoalescingConfig,
-}
-
-impl CoalescingState {
-    /// Creates a new coalescing state with the given configuration.
-    #[must_use]
-    pub fn new(config: CoalescingConfig) -> Self {
-        Self {
-            pending_count: AtomicU32::new(0),
-            timer_armed: AtomicBool::new(false),
-            last_armed: Mutex::new(None),
-            config,
-        }
-    }
-
-    /// Record a pending interrupt.
-    ///
-    /// Returns `true` if immediate delivery is needed (count exceeds threshold).
-    pub fn record(&self) -> bool {
-        let count = self.pending_count.fetch_add(1, Ordering::Relaxed);
-        if count + 1 >= self.config.max_coalesce_count {
-            return true;
-        }
-        if count == 0 {
-            // First interrupt in window — arm timer
-            self.timer_armed.store(true, Ordering::Release);
-            *self.last_armed.lock().unwrap_or_else(|e| e.into_inner()) = Some(Instant::now());
-        }
-        false
-    }
-
-    /// Flush coalesced state. Returns the number of coalesced interrupts.
-    pub fn flush(&self) -> u32 {
-        let count = self.pending_count.swap(0, Ordering::SeqCst);
-        self.timer_armed.store(false, Ordering::Release);
-        *self.last_armed.lock().unwrap_or_else(|e| e.into_inner()) = None;
-        count
-    }
-
-    /// Check if the coalescing timer has expired.
-    #[must_use]
-    pub fn timer_expired(&self) -> bool {
-        if !self.timer_armed.load(Ordering::Acquire) {
-            return false;
-        }
-        let guard = self.last_armed.lock().unwrap_or_else(|e| e.into_inner());
-        guard.is_some_and(|armed_at| armed_at.elapsed() >= self.config.max_delay)
-    }
-}
-
-/// Statistics for interrupt coalescing.
-#[derive(Debug, Default)]
-pub struct IrqStats {
-    /// Total interrupts triggered.
-    pub triggered: AtomicU64,
-    /// Interrupts coalesced (not delivered because pending).
-    pub coalesced: AtomicU64,
-}
+use super::{
+    CoalescingConfig, CoalescingState, Gsi, Irq, IrqConfig, IrqStats, IrqTriggerCallback, MAX_GSIS,
+    MAX_IRQS, TriggerMode,
+};
 
 /// IRQ chip abstraction.
 ///
@@ -665,6 +477,7 @@ impl IrqChip {
 mod tests {
     use super::*;
     use std::sync::atomic::AtomicUsize;
+    use std::time::Duration;
 
     #[test]
     fn test_irq_allocation() {
@@ -843,7 +656,7 @@ mod tests {
             let mut configs = chip.irq_configs.write().unwrap_or_else(|e| e.into_inner());
             configs.insert(
                 legacy_irq,
-                super::IrqConfig {
+                IrqConfig {
                     gsi: 5,
                     trigger_mode: TriggerMode::Edge,
                     asserted: false,
@@ -871,10 +684,6 @@ mod tests {
         // These IRQs are never masked (is_masked returns false)
         assert!(!chip.is_masked(irq));
     }
-
-    // ==========================================================================
-    // CoalescingConfig / CoalescingState Tests
-    // ==========================================================================
 
     #[test]
     fn test_coalescing_config_presets() {
